extract mouse paw mode into a separate module
