+++ meson.build
@@ -13,6 +13,7 @@ sources = files([
   'src/mania.cpp',
   'src/mouse.cpp',
   'src/osu.cpp',
+  'src/mousepaw.cpp',
   'src/system.cpp',
   'src/taiko.cpp',
   'src/cats.cpp',
