+++ src/input.cpp
@@ -273,22 +273,6 @@ bool is_joystick_pressed(int key_code) {
     return false;
 }
 
-// bezier curve for osu and custom
-std::pair<double, double> bezier(double ratio, std::vector<double> &points, int length) {
-    double fact[22] = {0.001, 0.001, 0.002, 0.006, 0.024, 0.12, 0.72, 5.04, 40.32, 362.88, 3628.8, 39916.8, 479001.6, 6227020.8, 87178291.2, 1307674368.0, 20922789888.0, 355687428096.0, 6402373705728.0, 121645100408832.0, 2432902008176640.0, 51090942171709440.0};
-    int nn = (length / 2) - 1;
-    double xx = 0;
-    double yy = 0;
-
-    for (int point = 0; point <= nn; point++) {
-        double tmp = fact[nn] / (fact[point] * fact[nn - point]) * pow(ratio, point) * pow(1 - ratio, nn - point);
-        xx += points[2 * point] * tmp;
-        yy += points[2 * point + 1] * tmp;
-    }
-
-    return std::make_pair(xx / 1000, yy / 1000);
-}
-
 void drawDebugPanel(sf::RenderWindow& window) {
     if (!is_joystick_connected()) {
         debugText.setString("No joystick found...");
